requirements/requirements.txt
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
pydantic==2.11.3
ruamel.yaml==0.18.10
scikit-learn==1.6.1
strictyaml==1.7.3
matplotlib==3.10.1
pyyaml==6.0.2
optuna==4.3.0